updated transpose implementation

+++ examples/pagerank-transpose.rs
@@ -126,7 +126,7 @@ fn transpose(filename: String, index: usize, peers: usize) -> (Vec<u32>, Vec<(u3
 
         let len = rev.len();
         rev[len-1].1 += 1;
-        reversed.push(s / peers);
+        reversed.push(s / peers as u32);
     }
 
     return (deg, rev, reversed);
